Join SST-HTR nodes
